pyproject.toml
[project]
name = "course-paper-5"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0,<7.0)",
    "python-decouple (>=3.8)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "ipython (>=9.8.0,<10.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "email-verification (>=0.4.6,<0.5.0)",
    "redis (>=7.1.0,<8.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-yasg (>=1.21.15,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-celery-beat (>=2.9.0,<3.0.0)",
    "django-cors-headers (>=4.0.0,<5.0.0)",
    "celery (>=5.3.0,<6.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "coverage (>=7.13.5,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "autoflake (>=2.3.3,<3.0.0)",
    "eventlet (>=0.40.4,<0.41.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
